feat: Rename package `pcs-bridge` -> `canonical-bridge`
