+++ packages/canonical-bridge/src/configs/chains.ts
