+++ packages/canonical-bridge/src/configs/index.ts
