+++ packages/canonical-bridge/src/modules/i18n/locales/en.ts
