+++ packages/canonical-bridge/src/modules/i18n/locales/index.ts
