+++ packages/canonical-bridge/src/theme/breakpoints.ts
@@ -1,0 +1,6 @@
+export const breakpoints = {
+  base: '0px',
+  sm: '576px',
+  md: '576px',
+  lg: '968px',
+}
