+++ packages/canonical-bridge/src/token-config/mainnet/cBridge/config.json
