+++ packages/canonical-bridge/src/token-config/mainnet/chains.ts
