+++ packages/canonical-bridge/src/token-config/mainnet/layerZero/config.json
