+++ packages/canonical-bridge/src/token-config/mainnet/meson/config.json
