+++ packages/canonical-bridge/src/token-config/mainnet/stargate/config.json
