+++ packages/canonical-bridge/src/views/GlobalStyle/index.ts
