+++ packages/canonical-bridge/tsconfig.json
