+++ packages/canonical-bridge/vite.config.ts
