--- packages/pcs-bridge/src/constants/index.ts
@@ -1,6 +1,0 @@
-export const DEBOUNCE_DELAY = 1500
-export const ESTIMATE_AMOUNT_DATA_RELOAD = 30000
-
-export const CLIENT_TIME_OUT = 60 * 1000
-
-export const APP_NAME = 'pcs-bridge'
