--- packages/pcs-bridge/src/theme/breakpoints.ts
@@ -1,13 +1,0 @@
-export const breakpoints = {
-  base: '0px',
-  sm: '576px',
-  md: '576px',
-  lg: '968px',
-  // base: '0px',
-  // sm: '375px',
-  // md: '768px',
-  // lg: '1280px',
-  // xl: '1440px',
-  // // Must always provide value for '2xl' if using pixels.
-  // '2xl': '1440px',
-}
